Added an updated Programme page
Modified existing programme.html to remove the previous programme and layout.
Added new programme images.
Created new slideshow for the programme images and modified the CSS to add the slideshow effects.
Added programme back to header.js to allow it to be accessed from the other pages.
Added slideshow.js for JavaScript controls.